+++ RELEASE_NOTES.md
@@ -38,6 +38,10 @@ instead of a linter flag.
 * ADDED: New meta-flag, `linter`, to manage the values of all the linter
 flags as a group.
 
+* IMPROVED: Report as compilation errors instead of compilation warnings
+`use_module/1-2` directives when not compiling a module as an object while
+using a backend that doesn't support modules.
+
 * IMPROVED: Report as compilation errors `export/1` and `reexport/2`
 directives when not compiling a module as an object.